Fixed zerg trying to make Protoss detectors
Added an extra if statement so that only the Protoss will try to make
observers

Made it so if you dont manually put in a build order in production
manager, it will mass hydralisks by default. I tryed making it mass
zerglings but no build order was created ingame, I imagine this was the
"bug" that was mentioned for zerg by David Churchhill. It dosent havent
a problem making some units like hydralisks and mutalisks, but I tryed
with sunkens, lurkers and zerglins and it bugged out.

The changes in UalbertaBot.vcsproj was just me trying to change
UalbertaBot.dll to MarvinBot.dll, it did not work out so I reverted it
back to the Original.