Flushed out the zerg strategy a bit
Implimented some of the features that the Protoss build had that the
Zerg did not. Features include, more flushed out Hatchery expansion and
worker production.

I also implimented zerg hydralisk upgrading but it has yet to show up in
the ingame build order. This could be another potential bug, ill have to
test with a few more games to make sure this is the case.