Improved StarcraftData for the Zerg
There was a few Units and Buildings that were not included for the Zerg
in StarcraftData, so for them to be usable in the StrategyManager I had
to push them into a vector. There is still a few upgrades im missing so
you'll have to include them there if you want specific upgrades not
allready included.

Also started building a second Zerg Build order to start to debug why it
wont build more then 4 Zerglings.