Implimented Basic lurker burrow unburrow :D
The micro needs some fine tuneing, but so far the lurkers have basic
burrowing skills.

There was a crashing issue that came up, I recompiled it and the
crashing stopped...not sure if was due to this my lurker implemtation or
a "random" occurrence....if you are running into crashing issue, trying
using an opening book thats not the lurker book.